Refactor Web project to follow coding guidelines.
