+++ CinemaApp/Views/_ViewImports.cshtml
@@ -1,3 +1,3 @@
 ﻿@using CinemaApp
-@using CinemaApp.Models
+@using CinemaApp.ViewModels
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
